Init polar sensors in factory
